mysql: correctly parse DSN from URL
